[기영]안드로이드, ios의 getVideoAttribute 함수의 결과에 파일 사이즈도 표시하도록 수정(네이티브)

+++ android/app/src/main/java/com/anilog2/PhotoListUtil.java
@@ -327,7 +327,7 @@ public class PhotoListUtil {
             return new int[]{width, height};
         }
 
-        static double getFileSize(Uri uri, Context context) {
+        static long getFileSize(Uri uri, Context context) {
             try {
                 ParcelFileDescriptor f = context.getContentResolver().openFileDescriptor(uri, "r");
                 return f.getStatSize();
